Dynamic wrapping of children/parents working

+++ ext/python3-parser/Python3LexerBase.cpp
@@ -96,7 +96,7 @@ antlr4::CommonToken* Python3LexerBase::createDedent() {
 int Python3LexerBase::getIndentationCount(std::string spaces) {
   int count = 0;
 
-  for (int i = 0; i < spaces.length(); i ++) {
+  for (size_t i = 0; i < spaces.length(); i ++) {
     switch (spaces.at(i)) {
       case '\t':
         count += 8 - (count % 8);
